updated: "components" folder renamed to "/app/ui"
